requirements.txt
# Core Framework
streamlit==1.28.0
numpy==1.24.3
scipy==1.11.0

# Audio Processing
librosa==0.10.0
soundfile==0.12.1
pyaudio==0.2.13

# Deep Learning (CPU optimized for Heroku)
torch==2.0.1+cpu
torchaudio==2.0.2
speechbrain==0.5.14

# Real-Time Transcription
assemblyai==0.33.0

# Speech-to-Text
openai-whisper==20230314

# Data Analysis
pandas==2.0.3
plotly==5.15.0
scikit-learn==1.3.0

# NLP
textblob==0.17.1
nltk==3.8.1

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0
requests==2.31.0
Pillow==10.0.0